requirements.txt
Django
mysql
djangorestframework
